+++ builder_dependencies.sh
@@ -82,7 +82,7 @@ if [ -f ~/."$shell"rc ]; then
     sed -i '/PLE ()/,/^}/d' ~/."$shell"rc
   fi
   echo -e "${GREEN}PLE Builder has been successfully installed on your system.${NC}"
-  echo -e "${PURPLE}To run PLE Builder, exit the terminal/termux, reopen it & type 'PLE'.${NC}"
+  echo -e "${PURPLE}To run PLE Builder, exit the terminal/termux, reopen it & type 'PLE' or 'ple'.${NC}"
 
   # Get the current working directory of the script
   pwd=$(pwd)
@@ -112,4 +112,16 @@ else
   sleep 3
   touch ~/."$shell"rc
   chmod +x builder_dependencies.sh && ./builder_dependencies.sh
+fi
+
+# Define secondary alias 'ple'
+# Check if the configuration file exists
+if [ -f ~/."$shell"rc ]; then
+  # If it exists, check if the alias ple=PLE exists in the file
+  if grep -q "alias ple=PLE" ~/."$shell"rc; then
+  # If it exists, delete the line that contains the alias
+    sed -i '/alias ple=PLE/d' ~/."$shell"rc
+  fi
+  # If it exists, append the alias to the end of the file
+  echo "alias ple=PLE" >> ~/."$shell"rc
 fi